In this stage, you will write a program that:
Reads a string of 9 symbols from the input and displays them to the user in a 3x3 grid. The grid can contain only X, O and _ symbols.
Outputs a line of dashes --------- above and below the grid, adds a pipe | symbol to the beginning and end of each line of the grid, and adds a space between all characters in the grid.
Examples
The greater-than symbol followed by a space (> ) represents the user input. Note that it's not part of the input.

Example 1:

> O_OXXO_XX
---------
| O _ O |
| X X O |
| _ X X |
---------
Example 2:

> OXO__X_OX
---------
| O X O |
| _ _ X |
| _ O X |
---------
Example 3:

> _XO__X___
---------
| _ X O |
| _ _ X |
| _ _ _ |
---------